requirements.txt
BeautifulSoup==3.2.1
Django==1.8.5
Pillow==2.7.0
Unidecode==0.04.18
argparse==1.3.0
beautifulsoup4==4.3.2
boto==2.36.0
csscompressor==0.9.3
distribute==0.6.38
dj-database-url==0.1.2
django-ace==1.0.2
django-appconf==1.0.1
django-batch-uploader==0.10.1
-e git+https://github.com/ninapavlich/django-carbon.git@c3ac30222ba957af7da6f8ae3eab6358132d4597#egg=django_carbon-dev
#django-ckeditor==4.4.8
django-ckeditor-improved==1.4
django-debug-toolbar==1.3.2
django-extensions==1.5.0
django-extra-views==0.6.5
django-grappelli==2.7.1
django-heroku-memcacheify==0.8
django-heroku-postgresify==0.3
django-herokuify==1.0.pre3
django-htmlmin==0.7.0
django-imagekit==3.2.5
django-imagekit-cropper==1.10
django-inline-wrestler==4.4
django-localflavor==1.1
django-object-actions==0.5.1
django-pylibmc==0.5.0
django-pylibmc-sasl==0.2.4
django-reversion==1.8.7
django-robots==1.1
django-share-me-share-me==1.0
django-storages==1.1.8
django-unsaved-changes==0.9
gunicorn==19.2.1
html5lib==1.0b3
httplib2==0.9.1
libsass==0.7.0
newrelic==2.44.0.36
pilkit==1.1.12
ply==3.4#3.6 is not compatible with latest slimit
psycopg2==2.5.4
pylibmc==1.4.1
pyparsing==2.0.3
scss==0.8.73
six==1.9.0
slimit==0.8.1
sqlparse==0.1.15
wsgiref==0.1.2